requirements.txt
playwright
pandas
openai
python-dotenv
tqdm
aiohttp
nest_asyncio
